+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - FLA export - XML formatting with blank lines on Java9+
 - [#1974] DefineBits image reading problem
 - [#1963] AS2 properly decompile/direct edit long classes
+- [#1977] AS3 Find usages - class and function usages, various fixes
 
 ## [18.3.5] - 2023-02-12
 ### Added
@@ -2965,6 +2966,7 @@ All notable changes to this project will be documented in this file.
 [#1973]: https://www.free-decompiler.com/flash/issues/1973
 [#1974]: https://www.free-decompiler.com/flash/issues/1974
 [#1963]: https://www.free-decompiler.com/flash/issues/1963
+[#1977]: https://www.free-decompiler.com/flash/issues/1977
 [#1959]: https://www.free-decompiler.com/flash/issues/1959
 [#1960]: https://www.free-decompiler.com/flash/issues/1960
 [#1964]: https://www.free-decompiler.com/flash/issues/1964